roles/docker/molecule/default/test_default.py
"""
Molecule tests for docker role using testinfra.
"""
import os
import pytest
import testinfra.utils.ansible_runner

testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
    os.environ['MOLECULE_INVENTORY_FILE']
).get_hosts('all')


def test_docker_is_installed(host):
    """Verify Docker is installed."""
    docker = host.package("docker-ce")
    assert docker.is_installed


def test_docker_service_running(host):
    """Verify Docker service is running and enabled."""
    docker_service = host.service("docker")
    assert docker_service.is_running
    assert docker_service.is_enabled


def test_docker_group_exists(host):
    """Verify docker group exists."""
    docker_group = host.group("docker")
    assert docker_group.exists


def test_docker_command_available(host):
    """Verify docker command is available."""
    docker_cmd = host.run("docker --version")
    assert docker_cmd.rc == 0
    assert "Docker version" in docker_cmd.stdout


def test_docker_compose_installed(host):
    """Verify docker compose plugin is installed."""
    compose_cmd = host.run("docker compose version")
    assert compose_cmd.rc == 0


def test_docker_info(host):
    """Verify docker info command works."""
    docker_info = host.run("docker info")
    assert docker_info.rc == 0
    assert "Server Version:" in docker_info.stdout


def test_docker_daemon_config(host):
    """Verify Docker daemon configuration file exists."""
    daemon_config = host.file("/etc/docker/daemon.json")
    assert daemon_config.exists
    assert daemon_config.is_file
    assert daemon_config.mode == 0o644


def test_docker_hello_world(host):
    """Test Docker functionality with hello-world container.
    
    Note: This test may fail on some architectures (e.g., ARM on x86 emulation)
    due to platform incompatibilities. This is not indicative of Docker being
    improperly installed.
    """
    result = host.run("docker run --rm hello-world")
    # Allow test to pass if Docker is functional, even if hello-world fails
    # due to architecture issues (common on macOS Docker Desktop with Rocky Linux)
    if result.rc != 0 and "invalid argument" in result.stderr:
        # Docker is working, but hello-world has architecture issues
        pytest.skip("hello-world container has architecture compatibility issues")
    assert result.rc == 0
    assert "Hello from Docker!" in result.stdout


def test_user_in_docker_group(host):
    """Verify ansible user is in docker group."""
    user = host.user("ansible")
    assert "docker" in user.groups


def test_docker_daemon_json_contains_insecure_registries(host):
    """Verify Docker daemon.json contains insecure-registries configuration."""
    import json
    
    daemon_config_file = host.file("/etc/docker/daemon.json")
    assert daemon_config_file.exists
    
    # Parse the JSON content
    daemon_config = json.loads(daemon_config_file.content_string)
    
    # Check that insecure-registries key exists
    assert "insecure-registries" in daemon_config
    assert isinstance(daemon_config["insecure-registries"], list)
    
    # Verify expected insecure registries are configured
    expected_registries = [
        "localhost:5000",
        "registry.test.local:5000",
        "192.168.100.100:5000"
    ]
    
    for registry in expected_registries:
        assert registry in daemon_config["insecure-registries"], \
            f"Registry {registry} not found in insecure-registries"


def test_docker_daemon_json_valid_format(host):
    """Verify Docker daemon.json is valid JSON format."""
    import json
    
    daemon_config_file = host.file("/etc/docker/daemon.json")
    
    try:
        json.loads(daemon_config_file.content_string)
    except json.JSONDecodeError as e:
        pytest.fail(f"Invalid JSON in daemon.json: {e}")


# ============================================================================
# PERFORMANCE OPTIMIZATION TESTS
# ============================================================================
# NOTE: These tests are for configuration validation only.
# In Docker-in-Docker (Molecule), storage-driver may be auto-selected (vfs)
# for compatibility, so we only verify IF configured, not enforce it.

def test_docker_storage_driver_optimized(host):
    """Verify Docker storage driver configuration (optional in DinD)."""
    import json
    
    daemon_config_file = host.file("/etc/docker/daemon.json")
    daemon_config = json.loads(daemon_config_file.content_string)
    
    # In DinD, storage-driver may not be set (auto-detect for compatibility)
    # Only validate if it's explicitly configured
    if "storage-driver" in daemon_config:
        assert daemon_config["storage-driver"] == "overlay2", \
            f"Expected overlay2, got {daemon_config['storage-driver']}"


def test_docker_info_storage_driver(host):
    """Verify Docker storage driver is functional (may be vfs in DinD)."""
    docker_info = host.run("docker info --format '{{.Driver}}'")
    
    assert docker_info.rc == 0
    driver = docker_info.stdout.strip()
    
    # In DinD, common drivers are: vfs (slow but compatible), overlay2 (fast)
    # We just verify Docker is working, don't enforce specific driver
    assert driver in ["overlay2", "vfs", "overlay"], \
        f"Docker using unexpected storage driver: {driver}"
    assert docker_info.rc == 0
    assert "overlay2" in docker_info.stdout.lower(), \
        f"Docker not using overlay2: {docker_info.stdout}"


def test_crun_available(host):
    """Verify crun is installed for better performance."""
    # crun installation is optional, so we just check if it exists
    crun_check = host.run("which crun || echo 'not found'")
    # Don't fail if not found, just report
    if "not found" not in crun_check.stdout:
        # If crun is found, verify it's executable
        crun_version = host.run("crun --version")
        assert crun_version.rc == 0, "crun found but not executable"


def test_docker_logging_configuration(host):
    """Verify Docker logging is properly configured."""
    import json
    
    daemon_config_file = host.file("/etc/docker/daemon.json")
    daemon_config = json.loads(daemon_config_file.content_string)
    
    assert "log-driver" in daemon_config
    assert daemon_config["log-driver"] == "json-file"
    
    assert "log-opts" in daemon_config
    log_opts = daemon_config["log-opts"]
    
    assert "max-size" in log_opts
    assert "max-file" in log_opts
    
    # Verify sensible defaults
    max_size = log_opts["max-size"]
    assert max_size.endswith("m") or max_size.endswith("M"), \
        "max-size should be in megabytes"
